Some win32 clean-ups.

+++ extras/the jucer/src/model/jucer_BinaryResources.cpp
@@ -261,7 +261,7 @@ void BinaryResources::loadFromCpp (const File& cppFileLocation, const String& cp
                         break;
                 }
 
-                jassert (size < out.getDataSize() && size > out.getDataSize() - 2);
+                jassert (size < (int) out.getDataSize() && size > (int) out.getDataSize() - 2);
 
                 MemoryBlock mb (out.getData(), out.getDataSize());
                 mb.setSize (size);
